backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.18
